# plugin/static/openapi.yaml
openapi: 3.0.1
info:
  title: Restaurant Search and Reservation Plugin
  description: "This plugin provides APIs for searching restaurants and making reservations."
  version: 'v1'
servers:
  - url: http://localhost:5000
paths:
  /search_restaurants:
    get:
      operationId: searchRestaurants
      summary: Search for restaurants
      parameters:
      - in: query
        name: query
        schema:
            type: string
        required: true
        description: The search query for restaurants.
      - in: query
        name: date
        schema:
            type: string
        required: true
        description: The reservation date.
      - in: query
        name: party_size
        schema:
            type: integer
        required: true
        description: The number of people in the party.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/searchRestaurantsResponse'
  /make_reservation:
    post:
      operationId: makeReservation
      summary: Make a reservation at a restaurant
      parameters:
      - in: query
        name: restaurant_id
        schema:
            type: string
        required: true
        description: The ID of the restaurant.
      - in: query
        name: date
        schema:
            type: string
        required: true
        description: The reservation date.
      - in: query
        name: time
        schema:
            type: string
        required: true
        description: The reservation time.
      - in: query
        name: party_size
        schema:
            type: integer
        required: true
        description: The number of people in the party.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/makeReservationResponse'
components:
  schemas:
    Restaurant:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        rating:
          type: number
          format: float
        price_range:
          type: string
        link:
          type: string
          format: uri
    searchRestaurantsResponse:
      type: object
      properties:
        restaurants:
          type: array
          items:
            $ref: '#/components/schemas/Restaurant'
          description: The list of restaurants.
    makeReservation:
      type: object
      properties:
        restaurant_id:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        party_size:
          type: integer
          format: int32
      required:
        - restaurant_id
        - date
        - time
        - party_size
    makeReservationResponse:
      type: object
      properties:
        restaurant_id:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        party_size:
          type: integer
          format: int32
        reservation_status:
          type: string
        confirmation_number:
          type: string